Split renderer/panes/code.tsx into code-parts/ mirroring projects-parts/
